output: enable plugin registration with event type.
This commit relies on changes being accepted in the fluent-bit repository. The goal is to enable event-type registration for go proxy plugins.
This change introduces a new function that will allow event-type setting in registration. This is done in order to not introduce breaking changes to current plugins.
This new function can be used instead of the default FLBPluginRegisterFunction to set the event_type to a different value (metrics (2)).

Signed-off-by: jmccormick7 <johnmccormick03711@gmail.com>
Signed-off-by: jmccormick7 <jcm258@case.edu>