input: add setting event_type to match proposed struct changes
This change just adds setting the empty value for event_type on input. Inputs do not set event type.
This is related to a prosposed change in the fluent-bit library to ensure that the struct is the same format
as the proposed change to enable output plugins to set their event type.

Signed-off-by: jmccormick7 <jcm258@case.edu>